# packages/api-types/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer": []
    }
  ],
  "paths": {
...
    "/api/v1/chat/read": {
      "post": {
        "tags": ["chat-controller"],
        "operationId": "markChatMessagesRead",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatReadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatStateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/messages": {
      "post": {
        "tags": ["chat-controller"],
        "operationId": "sendChatMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatSendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/messages/{messageId}/reaction": {
      "post": {
        "tags": ["chat-controller"],
        "operationId": "reactToChatMessage",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatReactionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["chat-controller"],
        "operationId": "removeChatMessageReaction",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/messages/media": {
      "post": {
        "tags": ["chat-controller"],
        "operationId": "uploadChatMediaMessage",
        "parameters": [
          {
            "name": "replyToMessageId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": ["files"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChatMessageResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/chat/partner-nickname": {
      "patch": {
        "tags": ["chat-controller"],
        "operationId": "updateChatPartnerNickname",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatPartnerNicknameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatThreadResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chat/messages/{messageId}": {
      "patch": {
        "tags": ["chat-controller"],
        "operationId": "editChatMessage",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatEditMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/chat/messages/{messageId}/me": {
      "delete": {
        "tags": ["chat-controller"],
        "operationId": "deleteChatMessageForMe",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/chat/messages/{messageId}/everyone": {
      "delete": {
        "tags": ["chat-controller"],
        "operationId": "deleteChatMessageForEveryone",
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessageResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ChatPresenceResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": ["ONLINE", "OFFLINE"]
          },
          "lastSeenAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ChatStateResponse": {
        "type": "object",
        "properties": {
          "partnerTyping": {
            "type": "boolean"
          },
          "partnerPresence": {
            "$ref": "#/components/schemas/ChatPresenceResponse"
          }
        }
      },
      "ChatReadRequest": {
        "type": "object",
        "properties": {
          "upToMessageId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["upToMessageId"]
      },
      "ChatSendMessageRequest": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["TEXT", "EMOJI", "GIF", "MEDIA"]
          },
          "body": {
            "type": "string"
          },
          "gifUrl": {
            "type": "string"
          },
          "gifProviderId": {
            "type": "string"
          },
          "replyToMessageId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["type"]
      },
      "ChatAttachmentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": ["IMAGE", "VIDEO"]
          },
          "url": {
            "type": "string"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "position": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ChatMessageResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "senderUserId": {
            "type": "string",
            "format": "uuid"
          },
          "viewerMessage": {
            "type": "boolean"
          },
          "type": {
            "type": "string",
            "enum": ["TEXT", "EMOJI", "GIF", "MEDIA", "BUB", "SAFE_NOTICE"]
          },
          "body": {
            "type": "string"
          },
          "gifUrl": {
            "type": "string"
          },
          "gifProviderId": {
            "type": "string"
          },
          "reply": {
            "$ref": "#/components/schemas/ChatReplyPreviewResponse"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "editedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedForEveryone": {
            "type": "boolean"
          },
          "deliveryState": {
            "type": "string",
            "enum": ["DELIVERED", "SEEN"]
          },
          "editable": {
            "type": "boolean"
          },
          "deletableForEveryone": {
            "type": "boolean"
          },
          "viewerReaction": {
            "type": "string"
          },
          "reactions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatReactionSummaryResponse"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatAttachmentResponse"
            }
          },
          "safeItemCount": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ChatReactionSummaryResponse": {
        "type": "object",
        "properties": {
          "reaction": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "ChatReplyPreviewResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "senderUserId": {
            "type": "string",
            "format": "uuid"
          },
          "snippet": {
            "type": "string"
          },
          "deletedForEveryone": {
            "type": "boolean"
          }
        }
      },
      "ChatReactionRequest": {
        "type": "object",
        "properties": {
          "reaction": {
            "type": "string",
            "minLength": 1
          }
        },
        "required": ["reaction"]
      },
      "ChatPartnerNicknameRequest": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          }
        }
      },
...
      "ChatEditMessageRequest": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 0
          }
        },
        "required": ["body"]
      },
...
      "ChatThreadResponse": {
        "type": "object",
        "properties": {
          "hasActiveTether": {
            "type": "boolean"
          },
          "tetherConnectionId": {
            "type": "string",
            "format": "uuid"
          },
          "partnerDisplayName": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ChatStateResponse"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChatMessageResponse"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}